feat: add option `sideNavLayout`
Add support for theme option `sideNavLayout`.  If set to `factored`,
the side-nav menu will group items that share a common path prefix.

For example, if the default layout for side-nav looked like this:

	/config        >
        /config/alert  >
        /config/bacnet >
        /config/modbus >
        /log           >

Then the factored side-nav would look like this:

	/config        >
        /log           >

and after expanding /config, you'd see:

	/config        v
          GET /config
          PUT /config
          …/alert      >
          …/bacnet     >
          …/modbus     >
        /log           >

Note that this commit also updates the styling to show a vertical line
to the left of all side-bar <ul> lists that are not at the root level.
This probably should be configurable.